split javascript file
