+++ src/app.py
@@ -21,9 +21,14 @@ app = Flask(
     static_folder="../static"       # 静的ファイルフォルダを指定
 )
 
-UPLOAD_FOLDER = './uploads'
+UPLOAD_FOLDER = "./uploads"
 os.makedirs(UPLOAD_FOLDER, exist_ok=True)
-app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
+app.config["UPLOAD_FOLDER"] = UPLOAD_FOLDER
+
+# request head type
+RQHEADER_REGISTER="register" 
+RQHEADER_LOGIN="login" 
+
 
 """
 - [POST] upload function
@@ -31,6 +36,7 @@ app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 @app.route('/login')
 def login():
     return render_template("login.html")
+
 
 # ルートエンドポイントにアクセスしたときの処理を定義
 @app.route('/')
@@ -41,15 +47,21 @@ def home():
     }
     return render_template("index.jinja", **data)
 
+
 # ユーザーの認証
 @app.route('/upload', methods=['POST'])
 def upload_image():
     try:
         # 画像データを受け取る
         data = request.json.get('image')
+        request_head = request.json.get('request_head')
         if not data:
             return "No image data", 400
 
+        if request_head == RQHEADER_REGISTER:
+            print("user tried to login register")
+        elif request_head == RQHEADER_LOGIN:
+            print("user tried to login")
         # Base64形式をデコード
         header, encoded = data.split(',', 1)
         image_data = base64.b64decode(encoded)
@@ -67,27 +79,6 @@ def upload_image():
         return "Image uploaded successfully", 200
     except Exception as e:
         print(e)
-        return str(e), 500
-
-# ユーザーの登録
-@app.route('/register_image', methods=['POST'])
-def register_image():
-    try:
-        # 画像データを受け取る
-        data = request.json.get('image')
-        if not data:
-            return "No image data", 400
-        # Base64形式をデコード
-        header, encoded = data.split(',', 1)
-        image_data = base64.b64decode(encoded)
-        # ファイル保存
-        file_hash_name = hashlib.sha256(image_data).hexdigest()
-        file_path = os.path.join(app.config['UPLOAD_FOLDER'], '%s.png' % file_hash_name)
-        with open(file_path, 'wb') as f:
-            f.write(image_data)
-
-        return "Image uploaded successfully", 200
-    except Exception as e:
         return str(e), 500
 
 @app.route('/register')
