+++ templates/login.html
@@ -13,7 +13,7 @@
     <button id="capture">Capture</button>
     <button id="upload" disabled>Upload</button>
 
-    <script src="/static/script.js"></script>
+    <script type="module" src="./static/login_script.js"></script>
 </body>
 </html>
