+++ templates/register.jinja
@@ -14,8 +14,7 @@
     <button id="upload" disabled>Upload</button>
     <h2>new user name</h2>
     <input type="text"></input>
-
-    <script src="/static/script.js"></script>
+    <script  type="module" src="./static/register_script.js"></script>
 </body>
 </html>
